requirements.txt
absl-py==0.7.0
astor==0.7.1
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.2
Django==2.1.5
django-ckeditor==5.6.1
django-cors-headers==2.2.0
django-extensions==2.0.7
django-filter==2.1.0
django-filters==0.2.1
django-js-asset==1.2.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-suit==0.2.26
djangorestframework==3.9.1
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
idna==2.6
ipython==7.2.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.2
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.8
Markdown==3.0.1
MarkupSafe==1.0
numpy==1.16.1
openapi-codec==1.3.2
parso==0.3.2
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.8
protobuf==3.6.1
psycopg2==2.7.4
ptyprocess==0.6.0
PyBindGen==0.18.0.post3+g3eade76
Pygments==2.3.1
pytz==2018.4
PyYAML==3.13
requests==2.18.4
scipy==1.2.0
simplejson==3.14.0
six==1.11.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
